+++ RELEASENOTES.md
@@ -82,7 +82,7 @@ Build modes can now be specified per project
 - **VerifyPRChanges** is used to verify whether a PR contains changes, which are not allowed from a fork.
 
 ### New Settings
-- **buildTimeoutInMinutes** (project setting) is used to define a timeout for building an AL-Go project.
+- **buildTimeout** (project setting) is used to define a timeout for building an AL-Go project.
 
 ## v2.4